chore: move multichain/address.ts from app to shared
The file is a pure utility consumed by both UI and background.
Moving it to shared/lib/multichain/address.ts lets UI import it
without crossing the architectural boundary, eliminating ~17
eslint-disable-next-line import-x/no-restricted-paths comments
(and their accompanying TODOs).
