+++ ui/helpers/utils/multichain/blockExplorer.ts
@@ -1,9 +1,7 @@
 import { KnownCaipNamespace, parseCaipChainId } from '@metamask/utils';
 import { getAccountLink } from '@metamask/etherscan-link';
 import type { MultichainNetwork } from '../../../selectors/multichain/networks';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { normalizeSafeAddress } from '../../../../app/scripts/lib/multichain/address';
+import { normalizeSafeAddress } from '../../../../shared/lib/multichain/address';
 import { MultichainProviderConfig } from '../../../../shared/constants/multichain/networks';
 import { formatBlockExplorerAddressUrl } from '../../../../shared/lib/multichain/networks';
