+++ ui/pages/bridge/hooks/useExternalAccountResolution.ts
@@ -1,7 +1,6 @@
 import { useMemo, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-// eslint-disable-next-line import-x/no-restricted-paths
-import { isEthAddress } from '../../../../app/scripts/lib/multichain/address';
+import { isEthAddress } from '../../../../shared/lib/multichain/address';
 import type { ExternalDestinationAccount } from '../prepare/types';
 import {
   getDomainResolutions,
