+++ ui/pages/remove-snap-account/snap-account-card.tsx
@@ -40,9 +40,7 @@ import { AccountNetworkIndicator } from '../../components/multichain/account-net
 import { getAccountLabels } from '../../helpers/utils/accounts';
 import { useFormatters } from '../../hooks/useFormatters';
 import { getCurrentCurrency } from '../../ducks/metamask/metamask';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { normalizeSafeAddress } from '../../../app/scripts/lib/multichain/address';
+import { normalizeSafeAddress } from '../../../shared/lib/multichain/address';
 
 // Component to display snap account information (avatar, address, account group name, balance, network indicator, and snap name)
 export const SnapAccountCard = ({
